mira-core/mira-ingest/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
httpx==0.27.*
Pillow>=10.0
python-multipart==0.0.12
pytest==8.*
pytest-asyncio==0.24.*
anyio[trio]==4.12.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
pdfplumber==0.11.9
beautifulsoup4>=4.12
apify-client>=1.7
